requirements.txt
streamlit
faiss-cpu
sentence-transformers
pypdf2
python-docx
pandas
requests